Refactor into query-mutation-ui structure
